Restringindo os tipos de arquivos que podem ser enviados (png, jpg, jpeg) / Adicionando a validação de dados na rota de listagem dos dados se um restaurante

+++ server/src/modules/items/services/UpdateItemService.ts
@@ -105,6 +105,15 @@ class UpdateItemService {
 
         // Updating and saving the new image file in storage
         if (image) {
+            // Check if image type file is valid
+            const imageNameDotSepared = image.split(/[\s.]+/);
+            const imageExtention = imageNameDotSepared[imageNameDotSepared.length -1];
+
+            if(imageExtention !== 'png' && imageExtention !== 'jpg' && imageExtention !== 'jpeg') {
+                // Cancel the operation
+                throw new AppError('O tipo do arquivo enviado é inválido.');
+            }
+
             // Delete old file from storage
             await this.storageProvider.deleteFile(itemToUpdate.image);
