+++ server/src/modules/restaurants/services/UpdateProfileDataService.ts
@@ -66,7 +66,17 @@ class UpdateProfileDataService {
             );
         }
 
+        // Saving updated logo file in storage
         if (logo !== restaurant.logo && logo !== '') {
+            // Check if logo type file is valid
+            const logoNameDotSepared = logo.split(/[\s.]+/);
+            const logoExtention = logoNameDotSepared[logoNameDotSepared.length -1];
+
+            if(logoExtention !== 'png' && logoExtention !== 'jpg' && logoExtention !== 'jpeg') {
+                // Cancel the operation
+                throw new AppError('O tipo do arquivo enviado é inválido.');
+            }
+
             // Delete old logo file from storage
             await this.storageProvider.deleteFile(restaurant.logo);
